# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prime Wave API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - in: query
          name: topicId
          schema: { type: integer }
        - in: query
          name: studentId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/submit:
    post:
      operationId: submitAssignment
      tags: [assignments]
      summary: Submit assignment for AI evaluation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /assignments/{id}/evaluation:
    get:
      operationId: getAssignmentEvaluation
      tags: [assignments]
      summary: Get AI evaluation for an assignment submission
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"

  # ─── PROJECTS ───────────────────────────────────────────────────────────────
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates for a student
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
      responses:
        "200":
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ─── PORTFOLIO ──────────────────────────────────────────────────────────────
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Register a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
  /dashboard/student/{studentId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard summary
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
...
  /dashboard/hr/{companyId}:
    get:
      operationId: getHrDashboard
      tags: [dashboard]
      summary: Get HR dashboard for a company
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: HR dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrDashboard"
components:
  schemas:
...
    SemesterProgress:
      type: object
      required: [semesterId, studentId, completedModules, totalModules, completedTopics, totalTopics, overallScore, status]
      properties:
        semesterId:        { type: integer }
        studentId:         { type: integer }
        completedModules:  { type: integer }
        totalModules:      { type: integer }
        completedTopics:   { type: integer }
        totalTopics:       { type: integer }
        overallScore:      { type: number }
        status:            { type: string }
        creditsEarned:     { type: ["integer", "null"] }

    # ─── TOPIC ────────────────────────────────────────────────────────────────
...
    Assignment:
      type: object
      required: [id, topicId, title, description, type, status, maxScore]
      properties:
        id:          { type: integer }
        topicId:     { type: integer }
        title:       { type: string }
        description: { type: string }
        type:        { type: string }
        status:      { type: string }
        maxScore:    { type: number }
        deadline:    { type: ["string", "null"] }
        attempts:    { type: integer }
        maxAttempts: { type: integer }
    SubmissionInput:
      type: object
      required: [studentId, content]
      properties:
        studentId: { type: integer }
        content:   { type: string }
        githubUrl: { type: string }
        notes:     { type: string }
    Submission:
      type: object
      required: [id, assignmentId, studentId, status, submittedAt]
      properties:
        id:           { type: integer }
        assignmentId: { type: integer }
        studentId:    { type: integer }
        content:      { type: string }
        githubUrl:    { type: ["string", "null"] }
        status:       { type: string }
        score:        { type: ["number", "null"] }
        submittedAt:  { type: string }
    Evaluation:
      type: object
      required: [id, submissionId, overallScore, passed]
      properties:
        id:             { type: integer }
        submissionId:   { type: integer }
        overallScore:   { type: number }
        passed:         { type: boolean }
        accuracy:       { type: number }
        logic:          { type: number }
        codeQuality:    { type: number }
        documentation:  { type: number }
        security:       { type: number }
        performance:    { type: number }
        bestPractices:  { type: number }
        feedback:       { type: string }
        weakAreas:
          type: array
          items: { type: string }
        improvementPlan: { type: ["string", "null"] }
        evaluatedAt:    { type: string }

    # ─── PROJECT ──────────────────────────────────────────────────────────────
...
    Certificate:
      type: object
      required: [id, studentId, type, title, issuedAt, creditsAwarded, verificationCode]
      properties:
        id:               { type: integer }
        studentId:        { type: integer }
        type:             { type: string }
        title:            { type: string }
        semesterNumber:   { type: ["integer", "null"] }
        issuedAt:         { type: string }
        creditsAwarded:   { type: integer }
        score:            { type: ["number", "null"] }
        verificationCode: { type: string }
        qrCodeUrl:        { type: ["string", "null"] }

    # ─── PORTFOLIO ────────────────────────────────────────────────────────────
...
    CandidateMatch:
      type: object
      required: [studentId, studentName, matchScore, ppsScore, creditsEarned, semesterNumber]
      properties:
        studentId:        { type: integer }
        studentName:      { type: string }
        avatar:           { type: ["string", "null"] }
        university:       { type: string }
        matchScore:       { type: number }
        ppsScore:         { type: number }
        creditsEarned:    { type: integer }
        semesterNumber:   { type: integer }
        industryReadiness: { type: number }
        skills:
          type: array
          items: { type: string }
        githubUrl:        { type: ["string", "null"] }
        portfolioUrl:     { type: ["string", "null"] }
        skillsMatchPct:   { type: number }
    Company:
      type: object
      required: [id, name, email, industry, status, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        email:       { type: string }
        industry:    { type: string }
        website:     { type: ["string", "null"] }
        logo:        { type: ["string", "null"] }
        status:      { type: string }
        jobsPosted:  { type: integer }
        hiredCount:  { type: integer }
        createdAt:   { type: string }
    CompanyInput:
      type: object
      required: [name, email, industry]
      properties:
        name:     { type: string }
        email:    { type: string }
        industry: { type: string }
        website:  { type: string }

    # ─── DASHBOARDS ──────────────────────────────────────────────────────────
    StudentDashboard:
      type: object
      required: [studentId, semesterNumber, creditsEarned, ppsScore, topicsCompleted, assignmentsPassed, streakDays, industryReadiness, recentActivity]
      properties:
        studentId:           { type: integer }
        semesterNumber:      { type: integer }
        creditsEarned:       { type: integer }
        ppsScore:            { type: number }
        topicsCompleted:     { type: integer }
        modulesCompleted:    { type: integer }
        assignmentsPassed:   { type: integer }
        streakDays:          { type: integer }
        industryReadiness:   { type: number }
        nextTopic:           { type: ["string", "null"] }
        nextDeadline:        { type: ["string", "null"] }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        semesterProgress:
          $ref: "#/components/schemas/SemesterProgress"
    ActivityItem:
      type: object
      required: [id, type, title, timestamp]
      properties:
        id:        { type: integer }
        type:      { type: string }
        title:     { type: string }
        detail:    { type: ["string", "null"] }
        score:     { type: ["number", "null"] }
        timestamp: { type: string }
...
    HrDashboard:
      type: object
      required: [companyId, activeJobs, totalApplicants, totalHired, avgMatchScore, recentCandidates]
      properties:
        companyId:        { type: integer }
        activeJobs:       { type: integer }
        totalApplicants:  { type: integer }
        totalHired:       { type: integer }
        avgMatchScore:    { type: number }
        recentCandidates:
          type: array
          items:
            $ref: "#/components/schemas/CandidateMatch"