# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Prime Wave API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  score?: number | null;
  timestamp: string;
}