# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Prime Wave API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  email: string;
  industry: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logo?: string | null;
  status: string;
  jobsPosted?: number;
  hiredCount?: number;
  createdAt: string;
}