# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Prime Wave API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Evaluation {
  id: number;
  submissionId: number;
  overallScore: number;
  passed: boolean;
  accuracy?: number;
  logic?: number;
  codeQuality?: number;
  documentation?: number;
  security?: number;
  performance?: number;
  bestPractices?: number;
  feedback?: string;
  weakAreas?: string[];
  /** @nullable */
  improvementPlan?: string | null;
  evaluatedAt?: string;
}