# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Prime Wave API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { SemesterProgress } from './semesterProgress';

export interface StudentDashboard {
  studentId: number;
  semesterNumber: number;
  creditsEarned: number;
  ppsScore: number;
  topicsCompleted: number;
  modulesCompleted?: number;
  assignmentsPassed: number;
  streakDays: number;
  industryReadiness: number;
  /** @nullable */
  nextTopic?: string | null;
  /** @nullable */
  nextDeadline?: string | null;
  recentActivity: ActivityItem[];
  semesterProgress?: SemesterProgress;
}